bug disappear; don't know why

+++ DeviceHost/BaseClasses/BaseTrainer.h
@@ -60,8 +60,8 @@ private:
 
 public:
 	double total_init_t;
-	double total_find_fea_t;
-	double total_split_t;
+	long total_find_fea_t;
+	long total_split_t;
 	double total_prune_t;
 };
